+++ src/main/java/com/teleclimb/rest/repositories/RoundRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RoundRepository extends JpaRepository<Round, Long> {
-    List<Round> findByCompetitionId(Long competitionId);
+public interface RoundRepository extends JpaRepository<Round, Integer> {
+    List<Round> findByCompetitionId(Integer competitionId);
 }
